pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "raajadharma-log"
version = "1.0.0"
description = "Immutable append-only YAML ledger CLI for the RaajaDharma Archery Club DAO"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.9.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
x = ["tweepy>=4.14.0"]
screenshot = ["playwright>=1.40.0", "Pillow>=10.0.0"]
pdf = ["weasyprint>=60.0"]
dev = ["pytest>=7.4.0", "pytest-mock>=3.12.0", "ruff>=0.1.0"]

[project.scripts]
raajadharma-log = "scripts.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
ignore = ["E501", "E402", "PLC0415", "BLE001"]
